Move doctesting code into separate module
The code is complicated enough that it deserves a module separate from
DocChecks. Also, in the future, doctesting will ideally happen in a
separate step from the other doc checks -- DocChecks will be reserved
for checking the unified doctree.
