+++ docs/src/lib/internals/doctests.md
@@ -1,0 +1,5 @@
+# DocTests
+
+```@autodocs
+Modules = [Documenter.DocTests]
+```
