+++ src/Documenter.jl
@@ -46,6 +46,7 @@ include("Documents.jl")
 include("Builder.jl")
 include("Expanders.jl")
 include("CrossReferences.jl")
+include("DocTests.jl")
 include("DocChecks.jl")
 include("Writers/Writers.jl")
 include("Deps.jl")
